backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.119.0
Pillow==12.0.0
pydantic==2.12.3
pydantic_settings==2.11.0
pytest==8.4.2
SQLAlchemy==2.0.44
supabase==2.22.0
transformers==4.56.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
psycopg2==2.9.11
python-multipart==0.0.20
torch==2.9.0+cpu

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
